RCPP-41 Add missing symbolic links
