+++ CMakeLists.txt
@@ -92,8 +92,19 @@ if(DEFINED CPPREALM_USE_UV)
     if (DEFINED USES_CONAN)
         find_package(libuv REQUIRED)
         target_link_libraries(cpprealm PUBLIC ${libuv_LIBRARIES})
-        target_compile_definitions(cpprealm PUBLIC CPPREALM_USE_UV=1)
+    else()
+        find_package(libuv CONFIG REQUIRED)
+        if(TARGET libuv::uv_a)
+            target_link_libraries(cpprealm PUBLIC libuv::uv_a)
+        elseif(TARGET libuv::uv)
+            target_link_libraries(cpprealm PUBLIC libuv::uv)
+        elseif(BUILD_SHARED_LIBS)
+            target_link_libraries(cpprealm PUBLIC uv)
+        else()
+            target_link_libraries(cpprealm PUBLIC uv_a)
+        endif()
     endif()
+    target_compile_definitions(cpprealm PUBLIC CPPREALM_USE_UV=1)
 endif()
 
 install(EXPORT "${PROJECT_NAME}Targets"
